Partial commit for ongoing tests
